split class_command_completer
