+++ assistant_bot/class_commands_handler.py
@@ -32,3 +32,6 @@ class Commands_Handler:
                 return f"[/red]**** Exception other: {e} [/red]"
         return wrapper
     
+    #interface 
+    def backup_data():
+        ...
